requirements.txt
pyautogui==0.9.54
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-quartz==11.0
pymsgbox==1.0.9
pytweening==1.2.0
pyscreeze==1.0.1
pygetwindow==0.0.9
mouseinfo==0.1.3
pyrect==0.2.0
rubicon-objc==0.5.0
pyperclip==1.9.0
pynput==1.8.1
watchdog==6.0.0
